added native builder

+++ scripts/generate.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+echo "1. Checking ENV"
+./check_env.sh
+RC="$?"
+if [ $RC -ne 0 ]; then
+    exit $RC
+fi
+
 cd ..
 rm -rf bin
 RC="$?"
